+++ src/contract/money/src/entrypoint/token_freeze_v1.rs
@@ -35,10 +35,10 @@ use crate::{
 /// `get_metadata` function for `Money::TokenFreezeV1`
 pub(crate) fn money_token_freeze_get_metadata_v1(
     _cid: ContractId,
-    call_idx: u8,
+    call_idx: usize,
     calls: Vec<DarkLeaf<ContractCall>>,
 ) -> Result<Vec<u8>, ContractError> {
-    let self_ = &calls[call_idx as usize].data;
+    let self_ = &calls[call_idx].data;
     let params: MoneyTokenFreezeParamsV1 = deserialize(&self_.data[1..])?;
 
     // Public inputs for the ZK proofs we have to verify
@@ -66,10 +66,10 @@ pub(crate) fn money_token_freeze_get_metadata_v1(
 /// `process_instruction` function for `Money::TokenFreezeV1`
 pub(crate) fn money_token_freeze_process_instruction_v1(
     cid: ContractId,
-    call_idx: u8,
+    call_idx: usize,
     calls: Vec<DarkLeaf<ContractCall>>,
 ) -> Result<Vec<u8>, ContractError> {
-    let self_ = &calls[call_idx as usize].data;
+    let self_ = &calls[call_idx].data;
     let params: MoneyTokenFreezeParamsV1 = deserialize(&self_.data[1..])?;
 
     // We just check if the mint was already frozen beforehand
